+++ src/main/java/RIKU/server/Service/AdminService.java
@@ -1,5 +1,6 @@
 package RIKU.server.Service;
 
+import RIKU.server.Dto.User.Response.ReadPacersResponse;
 import RIKU.server.Dto.User.Response.ReadUsersResponse;
 import RIKU.server.Dto.User.Request.UpdateUserRoleRequest;
 import RIKU.server.Entity.Base.BaseStatus;
@@ -9,7 +10,6 @@ import RIKU.server.Security.AuthMember;
 import RIKU.server.Util.BaseResponseStatus;
 import RIKU.server.Util.Exception.Domain.UserException;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -17,7 +17,6 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
-@Slf4j
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
 public class AdminService {
@@ -49,5 +48,18 @@ public class AdminService {
 
             user.updateUserRole(request.getUserRole());
         });
+    }
+
+    // 페이서 조회
+    public List<ReadPacersResponse> getPacers(AuthMember authMember) {
+        // 운영진 권한 검증
+        if(!authMember.isAdmin()) {
+            throw new UserException(BaseResponseStatus.UNAUTHORIZED_USER);
+        }
+
+        return userRepository.findByIsPacer(Boolean.TRUE)
+                .stream()
+                .map(ReadPacersResponse::of)
+                .collect(Collectors.toList());
     }
 }
