+++ src/main/java/RIKU/server/Service/UserService.java
@@ -2,11 +2,9 @@ package RIKU.server.Service;
 
 import RIKU.server.Dto.User.Request.UpdateProfileRequest;
 import RIKU.server.Dto.User.Request.SignUpUserRequest;
-import RIKU.server.Dto.User.Response.ReadPacersResponse;
 import RIKU.server.Dto.User.Response.ReadUserProfileResponse;
 import RIKU.server.Entity.User.User;
 import RIKU.server.Repository.UserRepository;
-import RIKU.server.Security.AuthMember;
 import RIKU.server.Util.BaseResponseStatus;
 import RIKU.server.Util.Exception.Domain.UserException;
 import lombok.RequiredArgsConstructor;
@@ -16,8 +14,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.io.IOException;
-import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 @Slf4j
@@ -87,18 +83,5 @@ public class UserService {
         user.updateProfile(phone, password, profileImageUrl);
 
         return user.getId();
-    }
-
-    // 페이서 조회
-    public List<ReadPacersResponse> getPacers(AuthMember authMember) {
-        // 운영진 권한 검증
-        if(!authMember.isAdmin()) {
-            throw new UserException(BaseResponseStatus.UNAUTHORIZED_USER);
-        }
-
-        return userRepository.findByIsPacer(Boolean.TRUE)
-                .stream()
-                .map(ReadPacersResponse::of)
-                .collect(Collectors.toList());
     }
 }
